More module cleanup.
